+++ data_collectors/tinkoff_invest_stream_py/configuration/settings.py
@@ -1,6 +1,6 @@
 from dataclasses import dataclass, field
 
-__all__ = ("DataCollectionSettings", "StockFigi", "StorageSettings")
+__all__ = ("DataCollectionSettings", "StockFigi", "StorageSettings", "WatcherSettings")
 
 
 @dataclass(eq=False, repr=True)
@@ -21,3 +21,8 @@ class StorageSettings:
     # All internal storage settings are represented as dict. A storage class have to parse it himself.
     # Here, we avoid any strong dependencies and obligations
     settings: dict = field(default_factory=dict)
+
+
+@dataclass(eq=False, repr=True)
+class WatcherSettings:
+    max_sec_api_silence: int